refactoring

+++ src/plugins/intel_cpu/src/nodes/kernels/aarch64/jit_uni_eltwise_generic.hpp
@@ -179,9 +179,9 @@ private:
 
         std::vector<size_t> in_idxs;
         std::vector<size_t> aux_idxs;
-        for (int i = 0; i < eltwise_emitter->get_inputs_num(); i++)
+        for (size_t i = 0; i < eltwise_emitter->get_inputs_num(); i++)
             in_idxs.push_back(i);
-        for (int i = 0; i < eltwise_emitter->aux_vecs_count(); i++)
+        for (size_t i = 0; i < eltwise_emitter->aux_vecs_count(); i++)
             aux_idxs.push_back(i);
 
         std::vector<size_t> out_idxs;
